Move Choices to own module, add documentation
